requirements.txt
streamlit
torch
torchvision
Pillow
stegano